requirements.txt
python-dotenv
etherscan-python
web3
discord
